[PATCH] fix the subarch build again after ACPI breakage
The patch to check "maxcpus=N" early broke by putting maxcpus (a
variable which is only exported by mpparse.c) into
parse_cmdline_early().

The fix is to make it depend on the correct CONFIG_ option. 

In the subarchitectures:

CONFIG_X86_SMP is the one that means "I want standard x86 smp code" and
that's what this should depend on.