[Bluetooth] Allocate the BCM203x URB buffer seperately
The buffer shares cachelines with other fields of the structure. On
architectures which require synchronisation before or after doing DMA
to or from a buffer, any access to such fields sharing cachelines can
corrupt the content of the buffer. The allocation primitives guarantee
that cachelines are not shared.

Noticed by Oliver Neukum <oliver@neukum.org>