[Bluetooth] Fix race in RX complete routine of the USB drivers
The test of the HCI_RUNNING flags bit in the RX complete routine should
be made while holding the completion lock. Otherwise there can be a race
on SMP systems in which RX complete finds the bit is set and on another
processor the close routine clears the bit, grabs and releases the lock
and then unlinks all the active URB's. Then the first processor acquires
the lock and resubmits the URB.

Patch from Alan Stern <stern@rowland.harvard.edu>