[PATCH] Use valid node number when unmapping x86 CPUs
From: Rusty Russell <rusty@rustcorp.com.au>

From:  colpatch@us.ibm.com

The cpu_2_node[] array for i386 is initialized to all 0's, meaning that
until modified at CPU bring-up, all CPUs are mapped to node 0.

When CPUs are brought online, they are mapped to the appropriate node by
various mechanisms, depending on the underlying hardware.

When we unmap CPUs (hotplug time), we should return the mapping for the CPU
that is going away to its original state, ie: 0.

When this code was initially submitted, the misguided poster (me) made the
mistake of putting a -1 in the cpu_2_node[] array for the CPU going away.

This patch fixes this mistake, and allows code to get a valid node number
for all valid CPU numbers.  This is important, because most (if not all)
callers do not error check the value returned by the cpu_to_node() macro,
and they should not have to.  The API specifies that a valid node number be
returned for any valid CPU number.