[PATCH] USB: Patch for Clie TH55 Support in visor kernel module
I just want to explain the changes I've made, since I tried to adhere to the
more organized structure in the 2.6 visor module to incorporate my changes.

I didn't want to pollute this with some ugly hack.

I created a new device id table with name clie_id_5_table to cleanly separate
the UX50/TH55 from the rest.

Along with this I added a new usb_serial_device_type which is essentially a
copy of handspring_device but has its own attach function.

This new attach function ( clie_5_attach ) will do the actual magic.

I think it is justified to introduce a new type ( which I called clie_5 )
since I suspect that Sony is intentionally changing the interface spec ( As
it has done so in the past ). I dare to predict that we will see more clie
devices coming this year which will require this new attachment procedure. At
the moment I am only aware of the UX50 and TH55 to implement that weird
configuration.