[PATCH] USB: Altsetting updates for USB media drivers
This patch implements the new altsetting regime for the drivers under
usb/media.  Not much needed to be changed.  I'm unable to test any of the
changes, but at least they compile all right (except that I didn't even
try to compile the pwc driver since it's marked BROKEN).

The stv680 and w9968cf drivers still include an assumption that they are
bound to interface number 0.  Since that the drivers are fairly tightly
linked to a specific kind of device I didn't try to change those
assumptions, but maybe they should be changed.