[ARM PATCH] 1824/1: guard against gcc not respecting local variable register assignment
Patch from Nicolas Pitre

The recent discovery of a problem with gcc not always respecting register
assignment for local variables to be used with inline assembly is worrisome.
If unnoticed, such problems can cover bugs that might prove hard to find
especially when the code is right but silently ignored.  This patch adds a
test where needed to have the assembler confirm our register selection and
fail the kernel build if the wrong registers are allocated.