[PATCH] PCI Hotplug: Move an often used while loop to an inline function
Walking through a pci_resource list and freeing all members is done a lot of
times in unload functions. This patch moves this to an inline function in
pciehp_core.c, pciehp_pci.c, shpchp_core.c and shpchp_pci.c. This shrinks the
code a lot (some 200 lines) and makes it much easier to read. Also adds some
__exit.