[PATCH] USB: LEGO USB Tower driver v0.95
here is the latest version 0.95 of the LEGO USB Tower driver against 2.6.6-rc3
which corrects a lot of problems in the version currently in the kernel,
most notably sleeping in interrupt context and improper locking.
Please apply.

It has been thoroughly tested with UHCI, OHCI and EHCI host controllers
using Lejos and NQC.  Firmware and program download, and with proper
modifications all communication protocols supported by Lejos work,
as well as firmware and program download and datalog upload in NQC.

Notes to application maintainers/protocol designers:

- Small modifications are needed in communication protocols because
  the tower tends to discard the first byte of transmissions.
  So for example LNP needs to send an extra byte like 0xff before
  the packet, and F7 handlers needs to cope with a lost 0x55.

- I suggest /dev/usb/legousbtower0 etc. as the standard device names.
  This puts it in the same place as the other USB devices and makes
  clear which driver is responsible for these devices.